% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{brodmann}
\alias{brodmann}
\title{Brodmann Areas Atlas (1909)}
\usage{
brodmann()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Brodmann areas cortical parcellation
with 39 regions per hemisphere. Contains both 2D polygon
geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices
for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
brodmann()
plot(brodmann())
}
\references{
Brodmann K (1909). Vergleichende
Lokalisationslehre der Grosshirnrinde. Leipzig: Johann
Ambrosius Barth.

Pijnenburg R, et al. (2021). Myelo- and cytoarchitectonic
microstructural and functional human cortical atlases
reconstructed in common MRI space. \emph{NeuroImage}, 239,
118274. \doi{10.1016/j.neuroimage.2021.118274}
}
\seealso{
Other ggseg_atlases: 
\code{\link{campbell}()},
\code{\link{economo}()},
\code{\link{flechsig}()},
\code{\link{kleist}()},
\code{\link{smith}()}

Other cortical_atlases: 
\code{\link{campbell}()},
\code{\link{economo}()},
\code{\link{flechsig}()},
\code{\link{kleist}()},
\code{\link{smith}()}

Other historical_atlases: 
\code{\link{campbell}()},
\code{\link{economo}()},
\code{\link{flechsig}()},
\code{\link{kleist}()},
\code{\link{smith}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
\concept{historical_atlases}
