% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{economo}
\alias{economo}
\title{Economo & Koskinas Cytoarchitectonic Atlas (1925)}
\usage{
economo()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas based on the historical Economo & Koskinas 1925
cytoarchitectonic parcellation with 15 regions per
hemisphere. Contains both 2D polygon geometry for
\code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices for
\code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
economo()
plot(economo())
}
\references{
C.F. von Economo, G.N. Koskinas; Die
Cytoarchitektonik Der Hirnrinde Des Erwachsenen Menschen;
J. Springer (1925)

Pijnenburg R, et al. (2021). Myelo- and cytoarchitectonic
microstructural and functional human cortical atlases
reconstructed in common MRI space. \emph{NeuroImage}, 239,
118274. \doi{10.1016/j.neuroimage.2021.118274}
}
\seealso{
Other ggseg_atlases: 
\code{\link{brodmann}()},
\code{\link{campbell}()},
\code{\link{flechsig}()},
\code{\link{kleist}()},
\code{\link{smith}()}

Other cortical_atlases: 
\code{\link{brodmann}()},
\code{\link{campbell}()},
\code{\link{flechsig}()},
\code{\link{kleist}()},
\code{\link{smith}()}

Other historical_atlases: 
\code{\link{brodmann}()},
\code{\link{campbell}()},
\code{\link{flechsig}()},
\code{\link{kleist}()},
\code{\link{smith}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
\concept{historical_atlases}
