% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegHistorical-package.R
\docType{package}
\name{ggsegHistorical-package}
\alias{ggsegHistorical}
\alias{ggsegHistorical-package}
\title{ggsegHistorical: Historical Brain Atlases for the 'ggseg' Ecosystem}
\description{
Six historical cortical brain atlases digitally reconstructed by Pijnenburg et al. (2021) \doi{10.1016/j.neuroimage.2021.118274} from the Dutch Connectome Lab. Includes Brodmann (1909), Campbell (1905), Economo & Koskinas (1925), Flechsig (1920), Kleist (1934), and Smith (1907). Each atlas provides a unified 'ggseg_atlas' object with both 2D polygon geometry and 3D vertex indices, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggseg/ggsegHistorical}
  \item Report bugs at \url{https://github.com/ggseg/ggsegHistorical/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
