% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{smith}
\alias{smith}
\title{Smith Cortical Atlas (1907)}
\usage{
smith()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Smith 1907 cortical parcellation.
Contains both 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}
and 3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
smith()
plot(smith())
}
\references{
G.E. Smith; A new topographical survey of the
human cerebral cortex, being an account of the
distribution of the anatomically distinct cortical areas
and their relationship to the cerebral sulci;
J. Anat. Physiol., 41 (Pt 4) (1907), p. 237

Pijnenburg R, et al. (2021). Myelo- and cytoarchitectonic
microstructural and functional human cortical atlases
reconstructed in common MRI space. \emph{NeuroImage}, 239,
118274. \doi{10.1016/j.neuroimage.2021.118274}
}
\seealso{
Other ggseg_atlases: 
\code{\link{brodmann}()},
\code{\link{campbell}()},
\code{\link{economo}()},
\code{\link{flechsig}()},
\code{\link{kleist}()}

Other cortical_atlases: 
\code{\link{brodmann}()},
\code{\link{campbell}()},
\code{\link{economo}()},
\code{\link{flechsig}()},
\code{\link{kleist}()}

Other historical_atlases: 
\code{\link{brodmann}()},
\code{\link{campbell}()},
\code{\link{economo}()},
\code{\link{flechsig}()},
\code{\link{kleist}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
\concept{historical_atlases}
